requirements.txt
altgraph==0.17.5
babel==2.17.0
charset-normalizer==3.4.4
customtkinter==5.2.2
darkdetect==0.8.0
packaging==25.0
pefile==2024.8.26
pillow==12.0.0
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
pypiwin32==223
pywin32==311
pywin32-ctypes==0.2.3
reportlab==4.4.7
setuptools==80.9.0
tkcalendar==1.6.1
